pyproject.toml
[project]
name = "stormbpmn-new-doc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocs-static-i18n>=1.2.0",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-meta-manager>=0.3.0",
    "pyyaml>=6.0",
]
